add desimeter.trig with some utility funcs (especially degrees-handling trig functions), and use them
